+++ update_translations.sh
@@ -3,4 +3,4 @@
 cd "$(dirname "$(type -p "$0")")"
 
 cd recipes
-../manage.py makemessages
+../manage.py makemessages -a